+++ lua/plugins/ai/avante.lua
@@ -1,4 +1,155 @@
-local cursor_provider = require('custom.cursor_agent').acp_provider()
+local cursor_provider = require('custom.cursor_agent').acp_provider(
+  require('custom.cursor_agent').MODEL_COMPOSER_25
+)
+
+local function avante_keys()
+  local api = function()
+    return require 'avante.api'
+  end
+  local avante = function()
+    return require 'avante'
+  end
+
+  return {
+    {
+      '<leader>aa',
+      function()
+        api().ask()
+      end,
+      desc = 'Avante ask',
+      mode = { 'n', 'v' },
+    },
+    {
+      '<leader>an',
+      function()
+        api().ask { new_chat = true }
+      end,
+      desc = 'Avante new chat',
+      mode = { 'n', 'v' },
+    },
+    {
+      '<leader>az',
+      function()
+        api().zen_mode()
+      end,
+      desc = 'Avante zen mode',
+      mode = { 'n', 'v' },
+    },
+    {
+      '<leader>ae',
+      function()
+        api().edit()
+      end,
+      desc = 'Avante edit selection',
+      mode = 'v',
+    },
+    {
+      '<leader>aS',
+      function()
+        api().stop()
+      end,
+      desc = 'Avante stop request',
+    },
+    {
+      '<leader>ar',
+      function()
+        api().refresh()
+      end,
+      desc = 'Avante refresh',
+    },
+    {
+      '<leader>af',
+      function()
+        api().focus()
+      end,
+      desc = 'Avante focus sidebar',
+    },
+    {
+      '<leader>at',
+      function()
+        avante().toggle_sidebar()
+      end,
+      desc = 'Avante toggle sidebar',
+    },
+    {
+      '<leader>ad',
+      function()
+        avante().toggle.debug()
+      end,
+      desc = 'Avante toggle debug',
+    },
+    {
+      '<leader>aC',
+      function()
+        avante().toggle.selection()
+      end,
+      desc = 'Avante toggle selection',
+    },
+    {
+      '<leader>as',
+      function()
+        avante().toggle.suggestion()
+      end,
+      desc = 'Avante toggle suggestions',
+    },
+    {
+      '<leader>aR',
+      function()
+        require('avante.repo_map').show()
+      end,
+      desc = 'Avante repo map',
+    },
+    {
+      '<leader>a?',
+      function()
+        api().select_model()
+      end,
+      desc = 'Avante select model',
+    },
+    {
+      '<leader>ah',
+      function()
+        api().select_history()
+      end,
+      desc = 'Avante chat history',
+    },
+    {
+      '<leader>aM',
+      function()
+        api().select_acp_model()
+      end,
+      desc = 'Avante ACP model',
+    },
+    {
+      '<leader>am',
+      function()
+        api().select_acp_mode()
+      end,
+      desc = 'Avante ACP mode',
+    },
+    {
+      '<leader>aB',
+      function()
+        api().add_buffer_files()
+      end,
+      desc = 'Avante add all buffers',
+    },
+    {
+      '<leader>ac',
+      function()
+        local sidebar = select(1, avante().get(false))
+        if not sidebar or not sidebar:is_open() then
+          avante().open_sidebar { ask = false }
+          sidebar = select(1, avante().get(false))
+        end
+        if sidebar and sidebar:is_open() then
+          sidebar.file_selector:add_current_buffer()
+        end
+      end,
+      desc = 'Avante add current file',
+    },
+  }
+end
 
 return {
   {
@@ -11,8 +162,9 @@ return {
     init = function()
       require('custom.patches.agentic_acp_transport').apply()
     end,
-    config = function()
+    config = function(_, opts)
       require('custom.patches.avante_acp_selector').apply()
+      require('avante').setup(opts)
     end,
     opts = {
       provider = 'cursor-acp',
@@ -26,7 +178,7 @@ return {
       },
       behaviour = {
         auto_suggestions = false,
-        auto_set_keymaps = true,
+        auto_set_keymaps = false,
         auto_approve_tool_permissions = true,
         acp_follow_agent_locations = true,
       },
@@ -38,6 +190,7 @@ return {
         },
       },
     },
+    keys = avante_keys(),
     dependencies = {
       'nvim-lua/plenary.nvim',
       'MunifTanjim/nui.nvim',